Refactor the crate into v1 and v2 modules (for future stivale1 compatibility)
